Fix to bugs with observers and screen not being redrawn on observee death. Fix to short packet position reporting to correctly report negative positions (i.e. where players get sent on death). Some more tidying up of todo list.

+++ src/short.c
@@ -706,8 +706,8 @@ handleVPlayer (unsigned char *sbuf)
         {                       /* S_P2 */
             struct player_s2_spacket *pa2 = (struct player_s2_spacket *) sbuf;
 
-            x = SCALE * ntohs (pa2->x);
-            y = SCALE * ntohs (pa2->y);
+            x = SCALE * (short) ntohs (pa2->x);
+            y = SCALE * (short) ntohs (pa2->y);
             new_flags (ntohl (pa2->flags), 0);
         }
         else
